+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (unreleased)
 
 - Changed `item_id` to `user_id` for `similar_users`
+- Changed warning to an error when `value` passed to `fit`
 - Removed dependency on `wilson_score` gem for `top_items`
 - Dropped support for Ruby < 2.6